加算、減算、スカラ倍、行列積でblasライブラリが使用可能な場合使用するようにします。

+++ src/main.cpp
@@ -14,7 +14,8 @@ int main() {
 #ifdef USE_OPENBLAS
 	std::cout << "Using OpenBLAS for matrix operations." << std::endl;
 #endif
-    using MatrixType = Matrix<int>;
+
+    using MatrixType = Matrix<float>;
     MatrixType::Container2D data1 = { {1, 2, 3}, {4, 5, 6}, {7, 8, 9} };
     MatrixType::Container2D data2 = { {9, 8, 7}, {6, 5, 4}, {3, 2, 1} };
